# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Lomash Wood API
  version: 1.0.0
  description: >
    REST API for the Lomash Wood Kitchen & Bedroom Design and Consultation Platform.
    All endpoints are versioned under /v1. Authentication uses Bearer JWT tokens
    issued by the auth-service.
  contact:
    name: Lomash Wood Engineering
    email: engineering@lomashwood.co.uk
  license:
    name: Proprietary
servers:
  - url: https://api.lomashwood.co.uk/v1
    description: Production
  - url: https://api-staging.lomashwood.co.uk/v1
    description: Staging
  - url: http://localhost:4000/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders for authenticated customer
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Orders]
      summary: Create a new order (checkout initiation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created with payment intent client secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────
  # PAYMENTS
  # ─────────────────────────────────────────────
  /payments/create-intent:
    post:
      tags: [Payments]
      summary: Create a Stripe PaymentIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentIntentRequest'
      responses:
        '201':
          description: PaymentIntent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ─────────────────────────────────────────────
  # CONTENT
  # ─────────────────────────────────────────────
...
  /webhooks/stripe:
    post:
      tags: [Webhooks]
      summary: Stripe webhook receiver
      security: []
      parameters:
        - name: stripe-signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw Stripe event payload
      responses:
        '200':
          description: Webhook received
        '400':
          description: Invalid signature
        '422':
          description: Unhandled event type

# ─────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Request body validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
  schemas:
...
    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: string
                format: uuid
              quantity:
                type: integer
                minimum: 1
    CreateOrderResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        clientSecret:
          type: string
        totalAmount:
          type: number
          format: float
        currency:
          type: string
          example: GBP
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, CONFIRMED, PROCESSING, COMPLETED, CANCELLED, FAILED]
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: float
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productTitle:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        totalPrice:
          type: number
          format: float
    OrderListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ── Payments ──────────────────────────────
    CreatePaymentIntentRequest:
      type: object
      required: [orderId]
      properties:
        orderId:
          type: string
          format: uuid
    PaymentIntentResponse:
      type: object
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
        amount:
          type: integer
          description: Amount in pence
        currency:
          type: string

    # ── Content ───────────────────────────────
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 422
        error:
          type: string
          example: Unprocessable Entity
        message:
          type: string
          example: Validation failed
        issues:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        requestId:
          type: string